requirements.txt
# CiRA FutureEdge Studio - Python Dependencies
# Python 3.8+ required

# ============================================
# UI Framework
# ============================================
customtkinter==5.2.2
darkdetect==0.8.0
pillow==10.2.0

# ============================================
# Data Processing
# ============================================
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0

# ============================================
# Feature Engineering
# ============================================
tsfresh==0.20.2
pysr==0.16.3
librosa==0.10.1  # Optional: for audio features
pywavelets==1.5.0

# ============================================
# Machine Learning
# ============================================
scikit-learn==1.3.2
pyod==1.1.2
imbalanced-learn==0.11.0

# ============================================
# Local LLM
# ============================================
llama-cpp-python==0.2.56
huggingface-hub==0.20.3

# ============================================
# Visualization
# ============================================
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1
kaleido==0.2.1  # For plotly static export

# ============================================
# Code Generation
# ============================================
jinja2==3.1.3
black==24.1.1  # Code formatter
autopep8==2.0.4

# ============================================
# Build System
# ============================================
cmake==3.28.1
ninja==1.11.1

# ============================================
# File Formats
# ============================================
h5py==3.10.0
pyarrow==14.0.2  # For Parquet
openpyxl==3.1.2  # Excel support
cbor2==5.7.1  # For Edge Impulse CBOR format

# ============================================
# Utilities
# ============================================
tqdm==4.66.1
loguru==0.7.2
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0

# ============================================
# Testing (for development)
# ============================================
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# ============================================
# Packaging
# ============================================
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0

# ============================================
# Optional: Performance
# ============================================
numba==0.58.1  # JIT compilation
bottleneck==1.3.7  # Faster pandas operations

# ============================================
# Platform-Specific
# ============================================
# Windows only
pywin32==306; sys_platform == 'win32'
